+++ javascript/src/jobProcessor.ts
@@ -3,6 +3,7 @@ import {
   ConnectionOptions,
   AckPolicy,
   connect,
+  ConsumerInfo,
   DeliverPolicy,
   DiscardPolicy,
   JsMsg,
@@ -11,23 +12,11 @@ import {
   RetentionPolicy,
   StorageType,
 } from 'nats'
-import { nanos, defer } from './util'
+import { nanos, defer, getNextBackoff, nanosToMs } from './util'
 import _debug from 'debug'
 import { Deferred, JobDef } from './types'
 
 const debug = _debug('nats-jobs')
-
-/**
- * Get the next backoff based on the redelivery count. If given
- * an array and no item exists for the attempt number use the last
- * backoff in the array.
- */
-const getNextBackoff = (backoff: number | number[], msg: JsMsg) => {
-  if (Array.isArray(backoff)) {
-    return backoff[msg.info.redeliveryCount - 1] || backoff.at(-1)
-  }
-  return backoff
-}
 
 const createStream = async (conn: NatsConnection, def: JobDef) => {
   const jsm = await conn.jetstreamManager()
@@ -42,12 +31,14 @@ const createStream = async (conn: NatsConnection, def: JobDef) => {
     discard: DiscardPolicy.Old,
     deny_delete: false,
     deny_purge: false,
-    ...def.streamConfig
+    ...def.streamConfig,
   }
   debug('stream config %O', config)
   // Add stream
   return jsm.streams.add(config)
 }
+
+const defaultAckWait = nanos('10s')
 
 const createConsumer = (conn: NatsConnection, def: JobDef) => {
   // Consumer config
@@ -55,10 +46,10 @@ const createConsumer = (conn: NatsConnection, def: JobDef) => {
     durable_name: `${def.stream}Consumer`,
     max_deliver: def.numAttempts ?? 5,
     ack_policy: AckPolicy.Explicit,
-    ack_wait: nanos('10s'),
+    ack_wait: defaultAckWait,
     deliver_policy: DeliverPolicy.All,
     replay_policy: ReplayPolicy.Instant,
-    ...def.consumerConfig
+    ...def.consumerConfig,
   }
   debug('consumer config %O', config)
   const js = conn.jetstream()
@@ -70,6 +61,22 @@ const createConsumer = (conn: NatsConnection, def: JobDef) => {
   })
 }
 
+const extendAckTimeoutThresholdFactor = 0.75
+/**
+ * Automatically extend the ack timeout by periodically telling NATS
+ * we're working.
+ */
+const extendAckTimeout = (ackWait: number, msg: JsMsg): NodeJS.Timer => {
+  const intervalMs = nanosToMs(ackWait) * extendAckTimeoutThresholdFactor
+  return setInterval(() => {
+    debug('extend ack - wait: %d msg: %O', ackWait, msg.info)
+    msg.working()
+  }, intervalMs)
+}
+
+/**
+ * Call perform for each message received on the stream.
+ */
 export const jobProcessor = async (opts?: ConnectionOptions) => {
   const conn = await connect(opts)
   const js = conn.jetstream()
@@ -82,15 +89,20 @@ export const jobProcessor = async (opts?: ConnectionOptions) => {
    * To gracefully shutdown see stop method.
    */
   const start = async (def: JobDef) => {
+    let extendAckTimer: NodeJS.Timer | undefined
     debug('job def %O', def)
     const pullInterval = def.pullInterval ?? ms('1s')
     const backoff = def.backoff ?? ms('1s')
     const batch = def.batch ?? 10
+    const autoExtendAckTimeout = def.autoExtendAckTimeout ?? true
     // Create stream
     // TODO: Maybe handle errors better
     await createStream(conn, def).catch()
     // Create pull consumer
     const ps = await createConsumer(conn, def)
+    // Get consumer info
+    const consumerInfo = await ps.consumerInfo()
+    const ackWait = consumerInfo.config.ack_wait || defaultAckWait
     // Pull messages from the consumer
     const run = () => {
       ps.pull({ batch, expires: pullInterval })
@@ -101,20 +113,27 @@ export const jobProcessor = async (opts?: ConnectionOptions) => {
     timer = setInterval(run, pullInterval)
     // Consume messages
     for await (const msg of ps) {
-      debug('received', msg.info)
+      debug('received %O', msg.info)
       deferred = defer()
+      if (autoExtendAckTimeout) {
+        extendAckTimer = extendAckTimeout(ackWait, msg)
+      }
+
       try {
         await def.perform(msg, { signal: abortController.signal, def, js })
         debug('completed')
         // Ack message
         await msg.ackAck()
       } catch (e) {
-        debug('failed', e)
+        debug('failed %O', e)
         const backoffMs = getNextBackoff(backoff, msg)
-        debug('next backoff ms', backoffMs)
+        debug('next backoff ms %d', backoffMs)
         // Negative ack message with backoff
         msg.nak(backoffMs)
       } finally {
+        if (extendAckTimer) {
+          clearInterval(extendAckTimer)
+        }
         deferred.done()
       }
       // Don't process any more messages if stopping