+++ javascript/src/util.ts
@@ -1,7 +1,9 @@
 import ms from 'ms'
 import { Deferred } from './types'
+import { JsMsg, Nanos } from 'nats'
 
 export const nanos = (x: string) => ms(x) * 1e6
+export const nanosToMs = (x: Nanos | number | undefined) => (x ? x / 1e6 : 0)
 
 export const expBackoff = (
   startMs: number,
@@ -27,3 +29,16 @@ export function defer<A>(): Deferred<A> {
     promise,
   }
 }
+
+/**
+ * Get the next backoff based on the redelivery count. If given
+ * an array and no item exists for the attempt number use the last
+ * backoff in the array.
+ */
+export const getNextBackoff = (backoff: number | number[], msg: JsMsg) => {
+  if (Array.isArray(backoff)) {
+    return backoff[msg.info.redeliveryCount - 1] || backoff.at(-1)
+  }
+  return backoff
+}
+